feat: [#47815] Add keeping subtree hopefully

See merge request rammearkitektur/os2mo-data-import-and-export!1179

+++ tools/subtreedeleter.py
@@ -110,15 +110,21 @@ class SubtreeDeleter:
         return await tqdm.gather(*(self.deleter(path, uuid) for uuid in uuids))
 
     async def run(
-        self, subtree_uuid: str, delete_functions: bool, keep_functions: List[str] = []
+        self,
+        subtree_uuid: str,
+        delete_functions: bool,
+        keep_functions: List[str] = [],
+        delete_subtree: bool = False,
     ):
         org_uuid = await self.get_org_uuid()
         tree = await self.get_tree(org_uuid)
         subtree = self.find_subtree(subtree_uuid, tree)
 
-        print("Deleting subtree for {}".format(subtree_uuid))
         unit_uuids = self.get_tree_uuids(subtree)
-        await self.delete_from_lora(unit_uuids, "organisation/organisationenhed")
+        if delete_subtree:
+            print("Deleting subtree for {}".format(subtree_uuid))
+            await self.delete_from_lora(unit_uuids, "organisation/organisationenhed")
+            print(f"Successfully deleted subtree: {subtree_uuid}")
 
         if delete_functions:
             print(
@@ -144,6 +150,7 @@ async def subtreedeleter_helper(
     org_unit_uuid: str,
     delete_functions: bool = False,
     keep_functions: List[str] = [],
+    delete_subtree: bool = False,
     connections: int = 4,
 ) -> None:
     token_settings = TokenSettings()
@@ -152,7 +159,10 @@ async def subtreedeleter_helper(
         session.headers.update(token_settings.get_headers())
         deleter = SubtreeDeleter(session, connections=connections)
         await deleter.run(
-            org_unit_uuid, delete_functions, keep_functions=keep_functions
+            org_unit_uuid,
+            delete_functions,
+            keep_functions=keep_functions,
+            delete_subtree=delete_subtree,
         )
 
 
@@ -174,21 +184,32 @@ async def subtreedeleter_helper(
     help="List of functions that should not be deleted",
 )
 @click.option(
+    "--delete-subtree",
+    default=False,
+    help="Specify whether to delete subtrees. If set, all org_units within the subtree will be deleted",
+)
+@click.option(
     "--connections",
     default=4,
     help="The amount of concurrent requests made to OS2mo",
 )
-def main(org_unit_uuid, delete_functions, keep, connections):
+def main(org_unit_uuid, delete_functions, keep, delete_subtree, connections):
     """Delete an organisational unit and all units below.
 
     Given the uuid of an org_unit this will delete the unit and all units below it.
     Optionally also deletes organisation functions such as engagements, KLE and addresses.
+    Optionally deletes the organisation unit and its subtrees. Default behaviour is to keep organisation units, and must be set to "False" for deletion.
+    if wished to delete, when calling the function.
     To delete all organisation functions, except certain type(s), add them with --keep.
     Example:
-        metacli tools/subtreedeleter.py --org-unit-uuid=c9b4c61f-1d38-5f6a-2c9e-d001e7cf6bd0 --delete-functions --keep=Leder --keep=KLE
+        metacli tools/subtreedeleter.py --org-unit-uuid=c9b4c61f-1d38-5f6a-2c9e-d001e7cf6bd0 --delete-functions --keep=Leder --keep=KLE --delete-subtree=True
     """
     subtreedeleter_helper(
-        org_unit_uuid, delete_functions, keep_functions=keep, connections=connections
+        org_unit_uuid,
+        delete_functions,
+        keep_functions=keep,
+        delete_subtree=delete_subtree,
+        connections=connections,
     )
